requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
gradio==4.44.1
chromadb==0.5.15
ollama==0.3.3
pypdf==4.3.1
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==10.4.0
python-multipart==0.0.12
pydantic==2.9.2
requests==2.32.3
numpy==1.26.4
# sentence-transformers pulls torch automatically.
# For Apple Silicon MPS: pip install torch==2.3.0 torchvision==0.18.0 (no extra index needed on macOS).
# For CUDA: pip install torch==2.3.0+cu121 --index-url https://download.pytorch.org/whl/cu121
# For CPU-only: pip install torch==2.3.0+cpu --index-url https://download.pytorch.org/whl/cpu
torch>=2.1.0
sentence-transformers==3.1.1
langchain==0.3.1
langchain-community==0.3.1
langchain-chroma==0.1.4
langchain-ollama==0.2.0
openpyxl==3.1.5
python-docx==1.1.2
pandas==2.2.3
aiofiles>=22.0,<24.0
httpx==0.27.2
